topic 0 - directory restructuring
