--- src/topic0/point2/ConnectionFactoryProducer.java
@@ -1,9 +1,0 @@
-package topic0.point2;
-
-public class ConnectionFactoryProducer {
-
-  public static AbstractConnectionFactory getConnectionFactory(boolean openSource) {
-    return openSource ? new OpenSourceConnectionFactory() : new PropietaryConnectionFactory();
-  }
-
-}
