--- src/topic0/point2/DbmsNotFoundException.java
@@ -1,11 +1,0 @@
-package topic0.point2;
-
-/**
- * Exception threw by Connection Factories when a Database Management System string name doesn't
- * match.
- */
-public class DbmsNotFoundException extends Exception {
-  public DbmsNotFoundException(String message) {
-    super(message);
-  }
-}
