--- src/topic0/point2/mockconnection/OracleConnection.java
@@ -1,18 +1,0 @@
-package topic0.point2.mockconnection;
-
-import topic0.point2.ConnectionInfo;
-
-/** Mock Oracle Database Connection. */
-public class OracleConnection extends Connection {
-
-  public OracleConnection(ConnectionInfo info) {
-    super(info);
-    System.out.println("Oracle Database Connection created! " + info);
-  }
-
-  @Override
-  public void tryToConnect() {
-    System.out.println("Trying to connect to the Oracle Database Database...");
-  }
-
-}
