--- src/topic0/point3/Connection.java
@@ -1,5 +1,0 @@
-package topic0.point3;
-
-public interface Connection {
-  void tryToConnect();
-}
