--- src/topic0/point4/Connection.java
@@ -1,5 +1,0 @@
-package topic0.point4;
-
-public interface Connection {
-  void tryToConnect();
-}
