--- src/topic0/point4/ConnectionBuilder.java
@@ -1,11 +1,0 @@
-package topic0.point4;
-
-
-public interface ConnectionBuilder {
-  Connection build();
-
-  ConnectionBuilder setServerName(String serverName);
-  ConnectionBuilder setDatabaseName(String databaseName);
-  ConnectionBuilder setUsername(String username);
-  ConnectionBuilder setPassword(String password);
-}
