--- src/topic0/point4/Test.java
@@ -1,11 +1,0 @@
-package topic0.point4;
-
-public class Test {
-
-  public static void main(String[] args) {
-    Connection connection = new Director(new ConcreteConnectionBuilder()).constructConnection();
-
-    connection.tryToConnect();
-  }
-
-}
