+++ topic-0/src/point1/DatabaseConnection.java
@@ -1,0 +1,24 @@
+package point1;
+
+/** Singleton class, implemented using enums, for a Database Connection. */
+enum DatabaseConnection {
+  INSTANCE;
+
+  private ConnectionInfo info;
+
+  private DatabaseConnection() {
+    System.out.println("Enums are great!!");
+  }
+
+  public void setInfo(ConnectionInfo info) {
+    this.info = info;
+  }
+
+  public ConnectionInfo getInfo() {
+    return this.info;
+  }
+
+  public void tryToConnect() {
+    System.out.println("INFO: " + info + ". Trying to Connect...");
+  }
+}
