+++ topic-0/src/point2/ConnectionFactoryProducer.java
@@ -1,0 +1,9 @@
+package point2;
+
+public class ConnectionFactoryProducer {
+
+  public static AbstractConnectionFactory getConnectionFactory(boolean openSource) {
+    return openSource ? new OpenSourceConnectionFactory() : new PropietaryConnectionFactory();
+  }
+
+}
