+++ topic-0/src/point2/mockconnection/DerbyConnection.java
@@ -1,0 +1,18 @@
+package point2.mockconnection;
+
+import point2.ConnectionInfo;
+
+/** Mock Apache Derby Connection. */
+public class DerbyConnection extends Connection {
+
+  public DerbyConnection(ConnectionInfo info) {
+    super(info);
+    System.out.println("Apache Derby Connection created! " + info);
+  }
+
+  @Override
+  public void tryToConnect() {
+    System.out.println("Trying to connect to the Derby Database...");
+  }
+
+}
