+++ topic-0/src/point2/mockconnection/PgConnection.java
@@ -1,0 +1,19 @@
+package point2.mockconnection;
+
+import point2.ConnectionInfo;
+
+/** Mock PostgreSQL Connection. */
+public class PgConnection extends Connection {
+
+  public PgConnection(ConnectionInfo info) {
+    super(info);
+    System.out.println("PostgreSQL Connection created! " + info);
+  }
+
+  @Override
+  public void tryToConnect() {
+    System.out.println("Trying to connect to the PostgreSQL Database...");
+  }
+
+
+}
