+++ topic-0/src/point3/Connection.java
@@ -1,0 +1,5 @@
+package point3;
+
+public interface Connection {
+  void tryToConnect();
+}
