+++ topic-0/src/point3/Test.java
@@ -1,0 +1,13 @@
+package point3;
+
+public class Test {
+
+  public static void main(String[] args) {
+    ProxyConnection pc = new ProxyConnection("The Best Server", "Things",
+        "user", "pass");
+
+    pc.tryToConnect();
+    pc.showInformation();
+  }
+
+}
