+++ topic-0/src/point4/Connection.java
@@ -1,0 +1,5 @@
+package point4;
+
+public interface Connection {
+  void tryToConnect();
+}
