+++ topic-0/src/point4/Test.java
@@ -1,0 +1,11 @@
+package point4;
+
+public class Test {
+
+  public static void main(String[] args) {
+    Connection connection = new Director(new ConcreteConnectionBuilder()).constructConnection();
+
+    connection.tryToConnect();
+  }
+
+}
